refactor(tests): split git tests to new file
